requirements.txt
# MindSpore Sign Language Recognition
# Requirements

# MindSpore - Huawei Deep Learning Framework
mindspore>=2.2.0

# MediaPipe - Landmark extraction
mediapipe>=0.10.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0
